refactor(pwa): decompose cloud sync settings route
Summary:
- Move cloud sync account settings rows and dialogs into focused
  component modules.
- Move cloud account cache and route helper logic out of the route.
- Keep auth result parsing next to the auth client and refresh Lingui
  catalog references.

Rationale:
- The cloud sync route was introduced as a 1,933-line route module in
  the reviewed range. Splitting route-local UI and pure helper code
  keeps the route under the 1k-line review bar without changing the
  auth or sync behavior.

Tests:
- vp install
- vp run check --fix
- vp run lingui:extract
- vp run check
- vp run test
- vp run build
